Yet another way to cast to 64 bit integer
`long long` is causing this on some platforms (namely, mine)

In file included from ./CellLimits.h:5:
./XlsCell.h:248:32: warning: 'long long' is a C++11 extension [-Wc++11-long-long]
        strs << std::fixed << (long long)cell_->d;
                               ^
1 warning generated.

+++ src/XlsCell.h
@@ -245,7 +245,7 @@ public:
       // examples: social security or student number
       double intpart;
       if (std::modf(cell_->d, &intpart) == 0.0) {
-        strs << std::fixed << (long long)cell_->d;
+        strs << std::fixed << (int64_t)cell_->d;
       } else {
         strs << cell_->d;
       }
